rename a private module
