+++ lapce-data/src/editor/commands/mod.rs
@@ -19,7 +19,7 @@ pub mod undo;
 pub mod indent_line;
 pub mod outdent_line;
 
-mod indent;
+mod indentation;
 
 /// This structure handles text editing commands.
 pub struct EditCommandFactory<'a> {
